requirements.txt
# ==============================================================================
# ENGAGE BOT - Python Dependencies
# ==============================================================================

# Core LangChain
langchain==0.1.0
langchain-core==0.1.10
langchain-openai==0.0.5

# LLM Providers (install what you need)
openai>=1.10.0
# anthropic>=0.8.0  # Uncomment if using Claude
# together>=0.2.0   # Uncomment if using Together AI

# HTTP Client
requests>=2.31.0
urllib3>=2.0.0

# Configuration
python-dotenv>=1.0.0

# Date/Time handling
tzdata>=2023.3  # For timezone support

# Optional: FastAPI (if building REST API)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0

# Optional: Async support
httpx>=0.26.0
aiohttp>=3.9.0

# Optional: Database (if storing conversations)
# sqlalchemy>=2.0.25
# psycopg2-binary>=2.9.9  # PostgreSQL
# pymongo>=4.6.1          # MongoDB

# Optional: Redis (for caching)
# redis>=5.0.1

# Optional: Monitoring & Logging
# sentry-sdk>=1.40.0
# langsmith>=0.0.75

# Development Dependencies
# pytest>=7.4.3
# pytest-asyncio>=0.21.0
# black>=23.12.0
# flake8>=7.0.0
# mypy>=1.8.0

# Testing
# pytest-cov>=4.1.0
# pytest-mock>=3.12.0